requirements.txt
asgiref==3.8.1
Django==5.1.4
django-allauth==65.3.1
django-appconf==1.0.6
django-bootstrap5==24.3
django-compressor==4.5.1
django-model-utils==5.0.0
platformshconfig==2.4.0
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.1
tzdata==2024.2
